pyproject.toml
[tool.poetry]
name = "skl-app"
version = "0.1.0"
description = "Skeleton App"
authors = ["Bayu Satria G"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
psycopg2 = "^2.9.9"
asyncpg = "^0.28.0"
greenlet = "^3.0.1"
sqlmodel = "0.0.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
